[Refactor] #40 - Adapter 통합하기
