+++ presentation/src/main/java/com/woowahan/banchan/ui/viewmodel/SoupDishBanchanViewModel.kt
@@ -2,9 +2,13 @@ package com.woowahan.banchan.ui.viewmodel
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.woowahan.banchan.util.FilterBanchanListUtil
+import com.woowahan.banchan.ui.dialog.CartItemInsertBottomSheet
+import com.woowahan.banchan.util.filterType
+import com.woowahan.banchan.util.getNewListApplyCartState
 import com.woowahan.domain.model.BanchanModel
 import com.woowahan.domain.usecase.FetchSoupDishBanchanUseCase
+import com.woowahan.domain.usecase.InsertCartItemUseCase
+import com.woowahan.domain.usecase.RemoveCartItemUseCase
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.MutableSharedFlow
 import kotlinx.coroutines.flow.MutableStateFlow
@@ -15,7 +19,9 @@ import javax.inject.Inject
 
 @HiltViewModel
 class SoupDishBanchanViewModel @Inject constructor(
-    private val fetchSoupDishBanchanUseCase: FetchSoupDishBanchanUseCase
+    private val fetchSoupDishBanchanUseCase: FetchSoupDishBanchanUseCase,
+    private val insertCartItemUseCase: InsertCartItemUseCase,
+    private val removeCartItemUseCase: RemoveCartItemUseCase
 ) : ViewModel() {
     private val _dataLoading: MutableStateFlow<Boolean> = MutableStateFlow(false)
     val dataLoading = _dataLoading.asStateFlow()
@@ -56,6 +62,56 @@ class SoupDishBanchanViewModel @Inject constructor(
         }
     }
 
+    val clickInsertCartButton: (BanchanModel, Boolean) -> (Unit) = { banchan, isCartItem ->
+        viewModelScope.launch {
+            when (isCartItem) {
+                true -> removeItemFromCart(banchan)
+                else -> {
+                    val dialog = CartItemInsertBottomSheet(banchan) { item, count ->
+                        insertItemsToCart(item, count)
+                    }
+                    _eventFlow.emit(UiEvent.ShowCartBottomSheet(dialog))
+                }
+            }
+        }
+    }
+
+    private fun removeItemFromCart(banchanModel: BanchanModel) {
+        viewModelScope.launch {
+            _dataLoading.emit(true)
+            removeCartItemUseCase.invoke(banchanModel.hash)
+                .onSuccess {
+                    defaultBanchans = defaultBanchans.getNewListApplyCartState(banchanModel, false)
+                    _banchans.value = _banchans.value.getNewListApplyCartState(banchanModel, false)
+                }.onFailure {
+                    it.printStackTrace()
+                    it.message?.let { message ->
+                        _eventFlow.emit(UiEvent.ShowToast(message))
+                    }
+                }.also {
+                    _dataLoading.emit(false)
+                }
+        }
+    }
+
+    private fun insertItemsToCart(banchanModel: BanchanModel, count: Int) {
+        viewModelScope.launch {
+            _dataLoading.emit(true)
+            insertCartItemUseCase.invoke(banchanModel, count)
+                .onSuccess {
+                    defaultBanchans = defaultBanchans.getNewListApplyCartState(banchanModel, true)
+                    _banchans.value = _banchans.value.getNewListApplyCartState(banchanModel, true)
+                }.onFailure {
+                    it.printStackTrace()
+                    it.message?.let { message ->
+                        _eventFlow.emit(UiEvent.ShowSnackBar(message))
+                    }
+                }.also {
+                    _dataLoading.emit(false)
+                }
+        }
+    }
+
     private fun filterBanchan(filterType: BanchanModel.FilterType) {
         viewModelScope.launch {
             if (filterType ==
@@ -64,7 +120,7 @@ class SoupDishBanchanViewModel @Inject constructor(
                 _banchans.value = defaultBanchans
             } else {
                 kotlin.runCatching {
-                    _banchans.value = FilterBanchanListUtil.filter(defaultBanchans, filterType)
+                    _banchans.value = defaultBanchans.filterType(filterType)
                 }.onFailure {
                     it.printStackTrace()
                     it.message?.let { message ->
@@ -100,5 +156,6 @@ class SoupDishBanchanViewModel @Inject constructor(
     sealed class UiEvent {
         data class ShowToast(val message: String) : UiEvent()
         data class ShowSnackBar(val message: String) : UiEvent()
+        data class ShowCartBottomSheet(val bottomSheet: CartItemInsertBottomSheet) : UiEvent()
     }
 }
