+++ presentation/src/main/res/layout/fragment_soup_dish_banchan.xml
@@ -11,7 +11,7 @@
 
         <variable
             name="adapter"
-            type="com.woowahan.banchan.ui.soupdishbanchan.SoupDishBanchanAdapter" />
+            type="com.woowahan.banchan.ui.adapter.DefaultBanchanAdapter" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
